requirements-dev.txt
--index-url https://download.pytorch.org/whl/cu121

torch
torchvision
torchaudio
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-groq
chromadb 
pypdf
python-docx
docx2txt
pandas
openpyxl
python-pptx
tqdm
python-dotenv
tiktoken
sentence-transformers
pydantic
unstructured
extract-msg
pywin32
langchain-ollama